Trying to get everything to work after rebasing my code into TopTaggingLPC

+++ Tools/derivedTupleVariables.h
@@ -1808,7 +1808,7 @@ namespace plotterFunctions
             int nIsoTrks; 
             if( tr.checkBranch("loose_isoTrksLVec") )
             {
-                nIsoTrks = AnaFunctions::countIsoTrks(tr.getVec<TLorentzVector>("loose_isoTrksLVec"), tr.getVec<double>("loose_isoTrks_iso"), tr.getVec<double>("loose_isoTrks_mtw"), tr.getVec<int>("loose_isoTrks_pdgId"));
+                nIsoTrks = AnaFunctions::countIsoTrks(tr.getVec<TLorentzVector>("loose_isoTrksLVec"), tr.getVec<float>("loose_isoTrks_iso"), tr.getVec<float>("loose_isoTrks_mtw"), tr.getVec<int>("loose_isoTrks_pdgId"));
             }
             else
             {
